# swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "paths": {
        "/v1/testcontroller/meetupjs": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "tags": [
                    "testcontroller"
                ]
            }
        },
        "/v1/testcontroller/queryparam/:id": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "description": ""
                    }
                ],
                "tags": [
                    "testcontroller"
                ]
            }
        },
        "/v1/testcontroller/testinguy/:name": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "description": ""
                    }
                ],
                "tags": [
                    "testcontroller"
                ]
            }
        },
        "/v1/testcontroller/getAction/:id/:id2/:id3": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id3",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "description": ""
                    },
                    {
                        "name": "id2",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "description": ""
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "description": ""
                    }
                ],
                "tags": [
                    "testcontroller"
                ]
            }
        },
        "/v1/testcontroller/getActionWithoutParam": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "tags": [
                    "testcontroller"
                ]
            }
        },
        "/v1/testcontroller/postAction/:id": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "description": ""
                    }
                ],
                "tags": [
                    "testcontroller"
                ]
            },
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "description": ""
                    }
                ],
                "tags": [
                    "testcontroller"
                ]
            }
        },
        "/v1/testcontroller/testHeaders": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "tags": [
                    "testcontroller"
                ]
            }
        },
        "/v1/testcontroller2/getAction": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "tags": [
                    "testcontroller2"
                ]
            }
        },
        "/v1/testcontroller2/postAction/:param1/:param2": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "param2",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "description": ""
                    },
                    {
                        "name": "param1",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "description": ""
                    }
                ],
                "tags": [
                    "testcontroller2"
                ]
            }
        },
        "/v1/user": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "tags": [
                    "user"
                ]
            }
        }
    },
    "swagger": "2.0"
}